# Speed up GPU tests by avoiding device reset