--- a/cuda/test/base/math.cu
+++ b/cuda/test/base/math.cu
@@ -47,6 +47,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "cuda/base/math.hpp"
 #include "cuda/base/types.hpp"
+#include "cuda/test/utils.hpp"
 
 
 namespace {