--- a/cuda/test/components/merging.cu
+++ b/cuda/test/components/merging.cu
@@ -47,6 +47,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 
 #include "cuda/components/cooperative_groups.cuh"
+#include "cuda/test/utils.hpp"
 
 
 namespace {