--- a/cuda/test/components/prefix_sum.cu
+++ b/cuda/test/components/prefix_sum.cu
@@ -44,6 +44,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/base/array.hpp>
 
 
+#include "cuda/test/utils.hpp"
+
+
 namespace {