--- a/cuda/test/components/sorting.cu
+++ b/cuda/test/components/sorting.cu
@@ -44,6 +44,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/base/executor.hpp>
 
 
+#include "cuda/test/utils.hpp"
+
+
 namespace {