--- a/cuda/test/stop/residual_norm_reduction_kernels.cpp
+++ b/cuda/test/stop/residual_norm_reduction_kernels.cpp
@@ -36,6 +36,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <gtest/gtest.h>
 
 
+#include "cuda/test/utils.hpp"
+
+
 namespace {