--- a/cuda/test/utils.hpp
+++ b/cuda/test/utils.hpp
@@ -30,8 +30,8 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************<GINKGO LICENSE>*******************************/
 
-#ifndef GKO_CUDA_TEST_NORESET_HPP_
-#define GKO_CUDA_TEST_NORESET_HPP_
+#ifndef GKO_CUDA_TEST_UTILS_HPP_
+#define GKO_CUDA_TEST_UTILS_HPP_
 
 
 #include <ginkgo/core/base/executor.hpp>
@@ -51,4 +51,4 @@ auto no_reset_exec =
 }  // namespace
 
 
-#endif  // GKO_CUDA_TEST_NORESET_HPP_
+#endif  // GKO_CUDA_TEST_UTILS_HPP_