--- a/hip/test/stop/criterion_kernels.hip.cpp
+++ b/hip/test/stop/criterion_kernels.hip.cpp
@@ -40,6 +40,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/base/types.hpp>
 
 
+#include "hip/test/utils.hip.hpp"
+
+
 namespace {