--- a/hip/test/stop/residual_norm_reduction_kernels.hip.cpp
+++ b/hip/test/stop/residual_norm_reduction_kernels.hip.cpp
@@ -36,6 +36,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <gtest/gtest.h>
 
 
+#include "hip/test/utils.hip.hpp"
+
+
 namespace {