--- a/hip/test/utils.hip.hpp
+++ b/hip/test/utils.hip.hpp
@@ -30,56 +30,25 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************<GINKGO LICENSE>*******************************/
 
-
-#include <core/test/utils/assertions.hpp>
-
-
-#include <gtest/gtest.h>
+#ifndef GKO_HIP_TEST_NORESET_HIP_HPP_
+#define GKO_HIP_TEST_NORESET_HIP_HPP_
 
 
-#include <ginkgo/core/matrix/csr.hpp>
-#include <ginkgo/core/matrix/dense.hpp>
+#include <ginkgo/core/base/executor.hpp>
 
 
-#include "hip/test/utils.hip.hpp"
+#include "core/test/utils.hpp"
 
 
 namespace {
 
 
-class MatricesNear : public ::testing::Test {
-protected:
-    void SetUp()
-    {
-        ASSERT_GT(gko::HipExecutor::get_num_devices(), 0);
-        ref = gko::ReferenceExecutor::create();
-        hip = gko::HipExecutor::create(0, ref);
-    }
-
-    void TearDown()
-    {
-        if (hip != nullptr) {
-            ASSERT_NO_THROW(hip->synchronize());
-        }
-    }
-
-    std::shared_ptr<gko::ReferenceExecutor> ref;
-    std::shared_ptr<const gko::HipExecutor> hip;
-};
-
-
-TEST_F(MatricesNear, CanPassHipMatrix)
-{
-    auto mtx = gko::initialize<gko::matrix::Dense<>>(
-        {{1.0, 2.0, 3.0}, {0.0, 4.0, 0.0}}, ref);
-    auto csr_ref = gko::matrix::Csr<>::create(ref);
-    csr_ref->copy_from(mtx.get());
-    auto csr_mtx = gko::matrix::Csr<>::create(hip);
-    csr_mtx->copy_from(std::move(csr_ref));
-
-    GKO_EXPECT_MTX_NEAR(csr_mtx, mtx, 0.0);
-    GKO_ASSERT_MTX_NEAR(csr_mtx, mtx, 0.0);
-}
+// prevent device reset after each test
+auto no_reset_exec =
+    gko::HipExecutor::create(0, gko::ReferenceExecutor::create());
 
 
 }  // namespace
+
+
+#endif  // GKO_HIP_TEST_NORESET_HIP_HPP_