At the moment, all of our GPU tests store a device executor that gets destroyed after each test, which in turn executes `cuda/hipDeviceReset`. This causes each of our tests to take at least 800ms on NVIDIA GPUs (less on AMD GPUs), independent of the actual kernel runtime. I think this is not necessary and can be fixed by instantiating a second device executor in a global variable (other, cleaner solutions welcome!)

Somewhat related to #461 

Before:
```
151: [==========] Running 24 tests from 1 test case.
151: [----------] Global test environment set-up.
151: [----------] 24 tests from Dense
151: [ RUN      ] Dense.SingleVectorHipScaleIsEquivalentToRef
151: [       OK ] Dense.SingleVectorHipScaleIsEquivalentToRef (1012 ms)
...
151: [ RUN      ] Dense.CalculateTotalColsIsEquivalentToRef
151: [       OK ] Dense.CalculateTotalColsIsEquivalentToRef (864 ms)
151: [----------] 24 tests from Dense (22293 ms total)
```
After:
```
151: [==========] Running 24 tests from 1 test case.
151: [----------] Global test environment set-up.
151: [----------] 24 tests from Dense
151: [ RUN      ] Dense.SingleVectorHipScaleIsEquivalentToRef
151: [       OK ] Dense.SingleVectorHipScaleIsEquivalentToRef (95 ms)
...
151: [ RUN      ] Dense.CalculateTotalColsIsEquivalentToRef
151: [       OK ] Dense.CalculateTotalColsIsEquivalentToRef (20 ms)
151: [----------] 24 tests from Dense (1783 ms total)
```