@thoasm 2020-02-27
What about creating a `cuda/test/utils.hpp` header (which includes `core/test/utils.hpp`), and do the creation of the executor there? That would at least hide the not so pleasant code inside a header and would allow us to improve / remove that in one line instead of in every test-file.
This would obviously require us changing all includes of `core/test/utils.hpp` to `cuda/test/utils.hpp`, but that is easy to automate (only needs to be done once).