@upsj 2020-03-02
@thoasm Good idea, I followed your suggestion and only had to patch a small compilation issue with `assert` and the comma operator in `constexpr` contexts for HIP.