@yhmtsai 2020-03-03
Does `return GKO_ASSERT(...), value;` -> `GKO_ASSERT(...); return value;`  not work?
Or it can not be constexpr?