@upsj 2020-03-03
In C++11, constexpr functions are only allowed to consist of a single return statement, so that unfortunately doesn't work. The expected behavior for `NDEBUG` would be that `assert` is replaced by something like `void(0)`, but this weird combination of includes seems to let some asserts vanish. I did not investigate this further yet, since the workaround works.